requirements.txt
fastapi
uvicorn[standard]
pillow
cos-python-sdk-v5